requirements.txt
Django==2.2.13
django-filter==2.3.0
djangorestframework==3.11.0
importlib-metadata==1.6.1
Markdown==3.2.2
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
psycopg2
django-cors-headers==3.4.0
djangorestframework-jwt==1.11.0
